rename memory_bool_index -> simple_bool_index
