+++ lib/segment/src/index/field_index/bool_index/mod.rs
@@ -1,1 +1,1 @@
-pub mod memory_bool_index;
+pub mod simple_bool_index;
